% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fitted_prclmm}
\alias{fitted_prclmm}
\title{A fitted PRC LMM}
\format{
A list comprising step 2 and step 3 as obtained
during the estimation of a PRC LMM
}
\usage{
data(fitted_prclmm)
}
\description{
This list contains a fitted PRC LMM, where the CBOCP is
computed using 50 cluster bootstrap samples. It is
used to reduce the computing time in the example of
the function \code{performance_prc}. The simulated dataset 
on which the model was fitted was landmarked at t = 2.
}
\examples{
data(fitted_prclmm)
ls(fitted_prclmm)
}
\references{
Signorelli, M. (2023). pencal: an R Package for the Dynamic 
Prediction of Survival with Many Longitudinal Predictors.
arXiv preprint: arXiv:2309.15600

Signorelli, M., Spitali, P., Al-Khalili Szigyarto, C, 
The MARK-MD Consortium, Tsonaka, R. (2021). 
Penalized regression calibration: a method for the prediction 
of survival outcomes using complex longitudinal and 
high-dimensional data. Statistics in Medicine.
DOI: 10.1002/sim.9178
}
\seealso{
\code{\link{performance_prc}}
}
\author{
Mirko Signorelli
}
\keyword{datasets}
