% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fitted_prcmlpmm}
\alias{fitted_prcmlpmm}
\title{A fitted PRC MLPMM}
\format{
A list comprising step 2 and step 3 as obtained
during the estimation of a PRC MLPMM
}
\usage{
data(fitted_prclmm)
}
\description{
This list contains a fitted PRC MLPMM. It is
used to reduce the computing time in the example of
the function \code{survpred_prcmlpmm}
}
\examples{
data(fitted_prcmlpmm)
ls(fitted_prcmlpmm)
}
\references{
Signorelli, M., Spitali, P., Al-Khalili Szigyarto, C, 
The MARK-MD Consortium, Tsonaka, R. (2021). 
Penalized regression calibration: a method for the prediction 
of survival outcomes using complex longitudinal and 
high-dimensional data. Statistics in Medicine, 40 (27), 6178-6196.
DOI: 10.1002/sim.9178
}
\seealso{
\code{\link{survpred_prcmlpmm}}
}
\author{
Mirko Signorelli
}
\keyword{datasets}
