% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmlpmm.R
\name{getmlpmm}
\alias{getmlpmm}
\title{Extract model fits from step 1 of PRC-LMM}
\usage{
getmlpmm(object, yname)
}
\arguments{
\item{object}{the output of `fit_lmms`}

\item{yname}{a character giving the name of one of the longitudinal
outcomes modelled within one of the MLPMM}
}
\value{
The model summary as returned by `summary.multlcmm`
}
\description{
Utility function to extract the MLPMM summaries from a model fit
obtained through `fit_mlpmms`
}
\references{
Signorelli, M. (2023). pencal: an R Package for the Dynamic 
Prediction of Survival with Many Longitudinal Predictors.
arXiv preprint: arXiv:2309.15600

Signorelli, M., Spitali, P., Al-Khalili Szigyarto, C, 
The MARK-MD Consortium, Tsonaka, R. (2021). 
Penalized regression calibration: a method for the prediction 
of survival outcomes using complex longitudinal and 
high-dimensional data. Statistics in Medicine, 40 (27), 6178-6196.
DOI: 10.1002/sim.9178
}
\seealso{
\code{\link{fit_mlpmms}} and \code{\link{summary.multlcmm}}
}
\author{
Mirko Signorelli
}
