% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_prclmm.R
\name{print.prclmm}
\alias{print.prclmm}
\title{Print method for PRC-LMM model fits}
\usage{
\method{print}{prclmm}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{prclmm}}

\item{digits}{number of digits at which the printed estimated regression
coefficients should be rounded (default is 4)}

\item{...}{additional arguments}
}
\value{
Summary information about the fitted PRC-LMM model
}
\description{
Print method for PRC-LMM model fits
}
\references{
Signorelli, M., Spitali, P., Tsonaka, R. (2021). Poisson-Tweedie 
mixed-effects model: a flexible approach for the analysis of longitudinal RNA-seq
data. Statistical Modelling, 21 (6), 520-545. URL: https://doi.org/10.1177/1471082X20936017
}
\seealso{
\code{\link{fit_prclmm}}, \code{\link{summary.prclmm}}
}
\author{
Mirko Signorelli
}
