% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_prclmm.R
\name{summary.prclmm}
\alias{summary.prclmm}
\title{Summary method for PRC-LMM model fits}
\usage{
\method{summary}{prclmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{prclmm}}

\item{...}{additional arguments}
}
\value{
An object of class `sprclmm`
}
\description{
Summary method for PRC-LMM model fits
}
\references{
Signorelli, M. (2023). pencal: an R Package for the Dynamic 
Prediction of Survival with Many Longitudinal Predictors.
arXiv preprint: arXiv:2309.15600

Signorelli, M., Spitali, P., Al-Khalili Szigyarto, C, 
The MARK-MD Consortium, Tsonaka, R. (2021). 
Penalized regression calibration: a method for the prediction 
of survival outcomes using complex longitudinal and 
high-dimensional data. Statistics in Medicine, 40 (27), 6178-6196.
DOI: 10.1002/sim.9178
}
\seealso{
\code{\link{fit_prclmm}}, \code{\link{print.prclmm}}
}
\author{
Mirko Signorelli
}
