% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_prcmlpmm.R
\name{summary.prcmlpmm}
\alias{summary.prcmlpmm}
\title{Summary method for PRC-MLPMM model fits}
\usage{
\method{summary}{prcmlpmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{prcmlpmm}}

\item{...}{additional arguments}
}
\value{
An object of class `sprcmlpmm`
}
\description{
Summary method for PRC-MLPMM model fits
}
\references{
Signorelli, M., Spitali, P., Tsonaka, R. (2021). Poisson-Tweedie 
mixed-effects model: a flexible approach for the analysis of longitudinal RNA-seq
data. Statistical Modelling, 21 (6), 520-545. URL: https://doi.org/10.1177/1471082X20936017
}
\seealso{
\code{\link{fit_prcmlpmm}}, \code{\link{print.prcmlpmm}}
}
\author{
Mirko Signorelli
}
