% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ranefs.R
\name{summary.ranefs}
\alias{summary.ranefs}
\title{Summary for step 2 of PRC}
\usage{
\method{summary}{ranefs}(object)
}
\arguments{
\item{object}{the output of `summarize_lmms` or `summarize_mlpmms`}
}
\value{
Information about number of predicted random effects and sample size
}
\description{
Summary function to extract basic descriptives from `summarize_lmms`
and `summarize_mlpmms`
}
\references{
Signorelli, M. (2023). pencal: an R Package for the Dynamic 
Prediction of Survival with Many Longitudinal Predictors.
arXiv preprint: arXiv:2309.15600

Signorelli, M., Spitali, P., Al-Khalili Szigyarto, C, 
The MARK-MD Consortium, Tsonaka, R. (2021). 
Penalized regression calibration: a method for the prediction 
of survival outcomes using complex longitudinal and 
high-dimensional data. Statistics in Medicine, 40 (27), 6178-6196.
DOI: 10.1002/sim.9178
}
\seealso{
\code{\link{fit_lmms}}
}
\author{
Mirko Signorelli
}
