% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_prclmm.R
\name{print.prclmm}
\alias{print.prclmm}
\title{Print method for PRC-LMM model fits}
\usage{
\method{print}{prclmm}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{prclmm}}

\item{digits}{number of digits at which the printed estimated regression
coefficients should be rounded (default is 4)}

\item{...}{additional arguments}
}
\value{
Summary information about the fitted PRC-LMM model
}
\description{
Print method for PRC-LMM model fits
}
\references{
Signorelli, M. (2023). pencal: an R Package for the Dynamic 
Prediction of Survival with Many Longitudinal Predictors.
arXiv preprint: arXiv:2309.15600

Signorelli, M., Spitali, P., Al-Khalili Szigyarto, C, 
The MARK-MD Consortium, Tsonaka, R. (2021). 
Penalized regression calibration: a method for the prediction 
of survival outcomes using complex longitudinal and 
high-dimensional data. Statistics in Medicine, 40 (27), 6178-6196.
DOI: 10.1002/sim.9178
}
\seealso{
\code{\link{fit_prclmm}}, \code{\link{summary.prclmm}}
}
\author{
Mirko Signorelli
}
