% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_prclmm.R
\name{summary.prclmm}
\alias{summary.prclmm}
\title{Summary method for PRC-LMM model fits}
\usage{
\method{summary}{prclmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{prclmm}}

\item{...}{additional arguments}
}
\value{
An object of class `sprclmm`
}
\description{
Summary method for PRC-LMM model fits
}
\references{
Signorelli, M., Spitali, P., Tsonaka, R. (2021). Poisson-Tweedie 
mixed-effects model: a flexible approach for the analysis of longitudinal RNA-seq
data. Statistical Modelling, 21 (6), 520-545. URL: https://doi.org/10.1177/1471082X20936017
}
\seealso{
\code{\link{fit_prclmm}}, \code{\link{print.prclmm}}
}
\author{
Mirko Signorelli
}
