% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_prcmlpmm.R
\name{summary.prcmlpmm}
\alias{summary.prcmlpmm}
\title{Summary method for PRC-MLPMM model fits}
\usage{
\method{summary}{prcmlpmm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{prcmlpmm}}

\item{...}{additional arguments}
}
\value{
An object of class `sprcmlpmm`
}
\description{
Summary method for PRC-MLPMM model fits
}
\references{
Signorelli, M. (2023). pencal: an R Package for the Dynamic 
Prediction of Survival with Many Longitudinal Predictors.
arXiv preprint: arXiv:2309.15600

Signorelli, M., Spitali, P., Al-Khalili Szigyarto, C, 
The MARK-MD Consortium, Tsonaka, R. (2021). 
Penalized regression calibration: a method for the prediction 
of survival outcomes using complex longitudinal and 
high-dimensional data. Statistics in Medicine, 40 (27), 6178-6196.
DOI: 10.1002/sim.9178
}
\seealso{
\code{\link{fit_prcmlpmm}}, \code{\link{print.prcmlpmm}}
}
\author{
Mirko Signorelli
}
